Reset project back to initial files
I believe I was over-engineering the project, without using OOP I find
it difficult to do all the necessary separation of concerns between
the architecture and the delivery mechanism. I will approach it in a
more procedural fashion and leave these architectural decisions for a
later stage or a future project revisit.